Adding 'Configure' options to CLI for File Sink (#2818)
## Why make this change?

- Fixes issue #2577 
- We want to allow the file sink properties to be configurable through
the CLI command `dab configure`.

## What is this change?

This change adds the file sink properties to the configure command and
allows the user to change those properties through the CLI. It also
ensures that the path property exists if file sink is enabled.
- `ConfigOptions.cs`: Adds file sink properties to CLI command so that
they can be configured by the user.
- `ConfigGenerator.cs`: Writes the file sink properties to the config
file, and errors out if the user tries to add an invalid value.
- `RuntimeConfigValidator.cs`: Validates that
`runtime.telemetry.file.path` is not empty or null if the file sink is
enabled.

## How was this tested?

- [ ] Integration Tests
- [X] Unit Tests

Added tests that ensure the configure commands and validation work.
Also refactored the validation tests that followed the same pattern.

## Sample Request(s)
--runtime.telemetry.file.enabled
--runtime.telemetry.file.path
--runtime.telemetry.file.rolling-interval
--runtime.telemetry.file.retained-file-count-limit
--runtime.telemetry.file.file-size-limit-bytes

---------

+++ src/Config/ObjectModel/FileSinkOptions.cs
@@ -60,10 +60,10 @@ public record FileSinkOptions
     /// <summary>
     /// File size limit in bytes before a new file needs to be created.
     /// </summary>
-    public int FileSizeLimitBytes { get; init; }
+    public long FileSizeLimitBytes { get; init; }
 
     [JsonConstructor]
-    public FileSinkOptions(bool? enabled = null, string? path = null, RollingInterval? rollingInterval = null, int? retainedFileCountLimit = null, int? fileSizeLimitBytes = null)
+    public FileSinkOptions(bool? enabled = null, string? path = null, RollingInterval? rollingInterval = null, int? retainedFileCountLimit = null, long? fileSizeLimitBytes = null)
     {
         if (enabled is not null)
         {
@@ -107,7 +107,7 @@ public record FileSinkOptions
 
         if (fileSizeLimitBytes is not null)
         {
-            FileSizeLimitBytes = (int)fileSizeLimitBytes;
+            FileSizeLimitBytes = (long)fileSizeLimitBytes;
             UserProvidedFileSizeLimitBytes = true;
         }
         else
